+++ deepchecks/vision/context.py
@@ -21,7 +21,7 @@ from deepchecks.core.errors import (DatasetValidationError, DeepchecksNotImpleme
                                     DeepchecksValueError, ModelValidationError, ValidationError)
 from deepchecks.utils.logger import get_logger
 from deepchecks.vision._shared_docs import docstrings
-from deepchecks.core.task_type import TaskType
+from deepchecks.vision.task_type import TaskType
 from deepchecks.vision.utils.vision_properties import STATIC_PROPERTIES_FORMAT
 from deepchecks.vision.vision_data import VisionData
